content/media/test/test_decoder_disable.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=448600
-->
<head>
  <title>Test for Bug 448600</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=448600">Mozilla Bug 448600</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<pre id="test">
<script type="application/javascript">

function filename(uri) {
  return uri.substr(uri.lastIndexOf("/")+1);
}

function e(id) {
  return document.getElementById(id);
}

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var prefService = Components.classes["@mozilla.org/preferences-service;1"]
               .getService(Components.interfaces.nsIPrefService);
var branch = prefService.getBranch("media.");
var gOldOggPref = branch.getBoolPref("ogg.enabled");
var gOldWavePref = branch.getBoolPref("wave.enabled");

branch.setBoolPref("ogg.enabled", false);
branch.setBoolPref("wave.enabled", false);

var gLoadError = new Object();

gLoadError['video1'] = 0; 
gLoadError['video2'] = 0;
gLoadError['video3'] = 0;

var gErrorCount = 0;

SimpleTest.waitForExplicitFinish();

function finishTest() {
  is(e('video1').currentSrc,
     "",
     'video1 currentSrc should be empty when there\'s no playable typed source children');
  is(filename(e('video2').currentSrc),
     filename(e('video2').src),
     'video2 currentSrc should match src');
  is(filename(e('video3').currentSrc),
     filename(e('video3').src),
     'video3 currentSrc should match src');

  is(gLoadError['video1'], 2, "Expect one error per invalid source child on video1");
  is(gLoadError['video2'], 1, "Expect one error on video2");
  is(gLoadError['video3'], 1, "Expect one error on video3");
  
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  branch.setBoolPref("ogg.enabled", gOldOggPref);
  branch.setBoolPref("wave.enabled", gOldWavePref);

  SimpleTest.finish();
}

function videoError(event, id) {
  event.stopPropagation();
  gLoadError[id]++;
  gErrorCount++;
  if (gErrorCount >= 4) {
    finishTest();
  }
}

</script>

<video id="video1" onerror="videoError(event, 'video1');">
  <source type="video/ogg" src="320x240.ogv"/>
  <source type="audio/wave" src="r11025_u8_c1.wav" id='s2'/>
</video>

<video id="video2" src="320x240.ogv" onerror="videoError(event, 'video2');"></video>
<video id="video3" src="r11025_u8_c1.wav" onerror="videoError(event, 'video3');"></video>

</pre>

</body>
</html>
